Move all chessboard components to separate directory
