+++ frontend/src/components/global/chessboard/ChessboardGrid.tsx
