+++ frontend/src/components/global/chessboard/Square.tsx
